+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.85.7
 
+- Node: export libsignal-net remote config keys as an array.